FEAT: more breaking up server and working on client control
