+++ client/index.html
@@ -7,7 +7,10 @@
     <title>Vite + TS</title>
   </head>
   <body>
-    <div id="app"></div>
+    <div id="app">
+        <input />
+        <button id="connect">connect</button>
+    </div>
     <script type="module" src="/src/main.ts"></script>
   </body>
 </html>
